requirements.txt
# BPMN XML Generator Requirements

# Required for web interface
flask>=2.3.0

# Environment variable loading
python-dotenv>=1.0.0

# Claude API
anthropic>=0.40.0

# Pinecone — vector DB + inference embeddings
pinecone>=3.0.0

# Optional: For better XML handling
# lxml>=4.9.0

# For Excel export
openpyxl>=3.1.0

# Document extraction (PDF and Word)
pdfplumber>=0.10.0
python-docx>=1.1.0

# flask-cors removed — CORS handled natively via before/after_request hooks
